+++ examples/sinatra/app.rb
@@ -3,7 +3,7 @@
 require "sinatra"
 require "sinatra/cookies"
 
-CABLE_URL = ENV["ANYCABLE"] ? "ws://localhost:9293/cable" : "/cable"
+CABLE_URL = ENV.fetch("CABLE_URL", "/cable")
 
 class App < Sinatra::Application # :nodoc:
   set :public_folder, "assets"