+++ examples/sinatra/config/environment.rb
@@ -8,11 +8,4 @@ require_relative "../chat"
 
 LiteCable.config.log_level = Logger::DEBUG
 
-if ENV["ANYCABLE"]
-  require "anycable"
-
-  # Turn AnyCable compatibility mode
-  LiteCable.anycable!
-
-  AnyCable.connection_factory = Chat::Connection
-end
+AnyCable.connection_factory = Chat::Connection