+++ lib/lite_cable/broadcast_adapters/base.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+# frozen_string_literal: true
+
+module LiteCable
+  module BroadcastAdapters
+    class Base
+      def initialize(**options)
+        @options = options
+      end
+
+      private
+
+      attr_reader :options
+    end
+  end
+end